requirements.txt
Django==1.9.6
djangorestframework==3.5.3